requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.3
idna==3.10
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
narwhals==1.48.1
numpy==2.3.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
protobuf==6.31.1
psycopg2-binary==2.9.10
pyaes==1.6.1
pyarrow==21.0.0
pyasn1==0.6.1
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.41
streamlit==1.47.1
streamlit-autorefresh==1.0.1
Telethon==1.40.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
